requirements.txt
pandas
numpy
scikit-learn
matplotlib
wordcloud
Pillow
requests
tqdm
torch
transformers
datasets